Upgrade to Python 3 and Django 2.1

+++ judge/views/blog.py
@@ -1,9 +1,9 @@
-from django.core.urlresolvers import reverse
 from django.db.models import Q, Max, Count
 from django.http import Http404
+from django.urls import reverse
 from django.utils import timezone
 from django.utils.functional import lazy
-from django.utils.translation import ugettext as _
+from django.utils.translation import gettext as _
 from django.views.generic import ListView
 
 from judge.comments import CommentedDetailView
@@ -47,10 +47,10 @@ class PostList(ListView):
                 context['has_clarifications'] = clarifications.count() > 0
                 context['clarifications'] = clarifications.order_by('-date')
 
-        context['user_count'] = lazy(Profile.objects.count, int, long)
-        context['problem_count'] = lazy(Problem.objects.filter(is_public=True).count, int, long)
-        context['submission_count'] = lazy(Submission.objects.count, int, long)
-        context['language_count'] = lazy(Language.objects.count, int, long)
+        context['user_count'] = lazy(Profile.objects.count, int, int)
+        context['problem_count'] = lazy(Problem.objects.filter(is_public=True).count, int, int)
+        context['submission_count'] = lazy(Submission.objects.count, int, int)
+        context['language_count'] = lazy(Language.objects.count, int, int)
 
         context['post_comment_counts'] = {
             int(page[2:]): count for page, count in
